All the changes stem from modifications to ClassMirror, MethodMirror.
Their methods no longer return other ClassMirror objects in response to getSuperClass()
or getInterfaces() or getExceptions(). Instead, they return String or String[]. Detector looks up those classes as necessary. This change avoids mirroring classes that we
KNOW to be non pausable, those in "java." or "javax." packages.

Weaving is also faster because the weaver stops looking for a pausable method up in the hierarchy as soon as it encounters a super class name in one of these nonpausable packages.