Fix for Issue #17 thanks to ted stockwell.
From the issue page:
"The kilim.analysis.ClassWeaver class maintains a static list of state classes that have been generated. Therefore, when used in Eclipse, if kilim.S_L is generated once when one eclipse project is built, then it will not be generated again in a second eclipse project that requires it.
"

	modified:   src/kilim/analysis/ClassWeaver.java