(1) Added support for stackmap attribute. (2) Bumped up the version to 1.0, because Kilim has been stable for a while.
Future versions of the JVM will make stackmaps mandatory if the classfile major version is >= 50. To enable stackmap support, we upgrade to version 4.x of ASM, which forces upon us a few changes:
1. ASM Visitors are now abstract classes, not interfaces
2. Use LabelNodes instead of Labels everywhere.

The second one deserves some explanation. ASM's MethodNode.InsnList uses the 'AbstractInsnNode.next' to thread together instructions. This means that an instruction cannot belong to two different lists, and instructions cannot be squirreled away and added later on (as used to be the case when MethodNode.instructions was an array). Second, Labels contain flags such as Debug (corresponding to line numbers) that cannot be reset. This means that such labels cannot be made targets at a later stage.
To accomodate this behavior of ASM, we stop dealing with Labels completely, and only deal with LabelNodes. When generating woven code, we strip the Labels off of LabelNodes to ensure that the labels from the original class file are not reused. The LabelNodes will generate new labels automatically.

        modified:   .classpath
        modified:   .gitignore
        modified:   .settings/org.eclipse.jdt.core.prefs
        modified:   build.sh
        modified:   build.xml
        modified:   docs/manual.html
        deleted:    libs/asm-all-2.2.3.jar
        new file:   libs/asm-all-4.1.jar
        modified:   src/kilim/Constants.java
        modified:   src/kilim/analysis/AsmDetector.java
        modified:   src/kilim/analysis/BasicBlock.java
        modified:   src/kilim/analysis/CallWeaver.java
        modified:   src/kilim/analysis/ClassFlow.java
        modified:   src/kilim/analysis/ClassWeaver.java
        modified:   src/kilim/analysis/MethodFlow.java
        modified:   src/kilim/analysis/MethodWeaver.java
        modified:   src/kilim/analysis/NopInsn.java
        modified:   src/kilim/mirrors/CachedClassMirrors.java
        modified:   src/kilim/mirrors/ClassMirror.java
        modified:   src/kilim/mirrors/Mirrors.java
        modified:   src/kilim/mirrors/RuntimeClassMirrors.java
        modified:   src/kilim/tools/Asm.java
        modified:   src/kilim/tools/DumpClass.java
        modified:   src/kilim/tools/FlowAnalyzer.java
        modified:   test.sh
        modified:   test/kilim/test/ex/ExJSR.j